requirements.txt
streamlit>=1.28.0
python-magic
markdownify
requests
pydantic
PyYAML
GitPython>=3.1.40
chromadb>=0.4.0
langchain>=0.0.300
langchain-ollama>=0.3.1
ollama>=0.1.7